change percentage back to float

+++ backend/src/monarch_py/datamodels/model.py
@@ -81,7 +81,7 @@ class Association(ConfiguredBaseModel):
     )
     has_count: Optional[int] = Field(None, description="""count of out of has_total representing a frequency""")
     has_total: Optional[int] = Field(None, description="""total, devided by has_count, representing a frequency""")
-    has_percentage: Optional[int] = Field(
+    has_percentage: Optional[float] = Field(
         None,
         description="""percentage, which may be calculated from has_count and has_total, as 100 * quotient or provided directly, rounded to the integer level""",
     )
@@ -262,7 +262,7 @@ class DirectionalAssociation(Association):
     )
     has_count: Optional[int] = Field(None, description="""count of out of has_total representing a frequency""")
     has_total: Optional[int] = Field(None, description="""total, devided by has_count, representing a frequency""")
-    has_percentage: Optional[int] = Field(
+    has_percentage: Optional[float] = Field(
         None,
         description="""percentage, which may be calculated from has_count and has_total, as 100 * quotient or provided directly, rounded to the integer level""",
     )
